+++ lib/an_lifecycle_viewmodel.dart
@@ -1,3 +1,4 @@
 library an_lifecycle_viewmodel;
 
 export 'src/view_model.dart';
+export 'src/view_model_ref.dart';